pyproject.toml
[project]
name = "ptagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.8.0",
    "playwright>=1.56.0",
    "pyyaml>=6.0.3",
]
